[CI]Style: Convert `test/` to ruff format(Batch #3) (#6744)
### What this PR does / why we need it?
| File Path |
| :--- |
| `tests/e2e/nightly/single_node/models/scripts/test_single_node.py` |
|
`tests/e2e/nightly/single_node/ops/multicard_ops_a2/test_matmul_allreduce_add_rmsnorm.py`
|
|
`tests/e2e/nightly/single_node/ops/multicard_ops_a3/test_dispatch_ffn_combine.py`
|
|
`tests/e2e/nightly/single_node/ops/multicard_ops_a3/test_dispatch_ffn_combine_bf16.py`
|
|
`tests/e2e/nightly/single_node/ops/multicard_ops_a3/test_dispatch_gmm_combine_decode.py`
|
|
`tests/e2e/nightly/single_node/ops/singlecard_ops/test_add_rms_norm_bias.py`
|
|
`tests/e2e/nightly/single_node/ops/singlecard_ops/test_apply_top_k_top_p_custom.py`
|
|
`tests/e2e/nightly/single_node/ops/singlecard_ops/test_batch_matmul_transpose.py`
|
| `tests/e2e/nightly/single_node/ops/singlecard_ops/test_bgmv_expand.py`
|
| `tests/e2e/nightly/single_node/ops/singlecard_ops/test_bgmv_shrink.py`
|
| `tests/e2e/nightly/single_node/ops/singlecard_ops/test_fused_moe.py` |
|
`tests/e2e/nightly/single_node/ops/singlecard_ops/test_gating_top_k_softmax.py`
|
|
`tests/e2e/nightly/single_node/ops/singlecard_ops/test_gmm_swiglu_quant_weight_nz_tensor_list.py`
|
|
`tests/e2e/nightly/single_node/ops/singlecard_ops/test_grouped_matmul_swiglu_quant.py`
|
|
`tests/e2e/nightly/single_node/ops/singlecard_ops/test_mla_preprocess.py`
|
|
`tests/e2e/nightly/single_node/ops/singlecard_ops/test_mla_preprocess_nq.py`
|
|
`tests/e2e/nightly/single_node/ops/singlecard_ops/test_mla_preprocess_qdown.py`
|
|
`tests/e2e/nightly/single_node/ops/singlecard_ops/test_moe_init_routing_custom.py`
|
|
`tests/e2e/nightly/single_node/ops/singlecard_ops/test_npu_moe_gating_top_k.py`
|
|
`tests/e2e/nightly/single_node/ops/singlecard_ops/test_transpose_kv_cache_by_block.py`
|
|
`tests/e2e/nightly/single_node/ops/singlecard_ops/test_vocabparallelembedding.py`
|
|
`tests/e2e/nightly/single_node/ops/singlecard_ops/triton/test_causal_conv1d.py`
|
|
`tests/e2e/nightly/single_node/ops/singlecard_ops/triton/test_chunk_gated_delta_rule.py`
|
|
`tests/e2e/nightly/single_node/ops/singlecard_ops/triton/test_fused_qkvzba_split_reshape_cat.py`
|
|
`tests/e2e/nightly/single_node/ops/singlecard_ops/triton/test_fused_sigmoid_gating_delta_rule.py`
|
|
`tests/e2e/nightly/single_node/ops/singlecard_ops/triton/test_l2norm.py`
|
|
`tests/e2e/nightly/single_node/ops/singlecard_ops/triton/test_mrope.py`
|
|
`tests/e2e/nightly/single_node/ops/singlecard_ops/triton/test_muls_add.py`
|
|
`tests/e2e/nightly/single_node/ops/singlecard_ops/triton/test_penality.py`
|
|
`tests/e2e/nightly/single_node/ops/singlecard_ops/triton/test_prepare_inputs_padded.py`
|
|
`tests/e2e/nightly/single_node/ops/singlecard_ops/triton/test_rejection_sample.py`
|
| `tests/e2e/nightly/single_node/ops/singlecard_ops/triton/test_rope.py`
|
|
`tests/e2e/nightly/single_node/ops/singlecard_ops/triton/test_split_qkv_rmsnorm_mrope.py`
|
|
`tests/e2e/nightly/single_node/ops/singlecard_ops/triton/test_split_qkv_rmsnorm_rope.py`
|

### Does this PR introduce _any_ user-facing change?

### How was this patch tested?

- vLLM version: v0.15.0
- vLLM main:
https://github.com/vllm-project/vllm/commit/9562912cead1f11e8540fb91306c5cbda66f0007

---------

+++ tests/e2e/nightly/single_node/ops/singlecard_ops/triton/test_split_qkv_tp_rmsnorm_rope.py
@@ -21,10 +21,8 @@ DEFAULT_RTOL = 5e-3
 
 
 def _build_rope(num_tokens, rotary_dim, dtype, device):
-    cos = torch.from_numpy(
-        np.random.uniform(0, 1, [num_tokens, rotary_dim // 2])).to(dtype).to(device)
-    sin = torch.from_numpy(
-        np.random.uniform(0, 1, [num_tokens, rotary_dim // 2])).to(dtype).to(device)
+    cos = torch.from_numpy(np.random.uniform(0, 1, [num_tokens, rotary_dim // 2])).to(dtype).to(device)
+    sin = torch.from_numpy(np.random.uniform(0, 1, [num_tokens, rotary_dim // 2])).to(dtype).to(device)
     return cos.contiguous(), sin.contiguous()
 
 
@@ -97,15 +95,12 @@ def _reference_impl(
     q_var = q_f32.pow(2).mean(dim=-1, keepdim=True)
     k_var = k_f32.pow(2).mean(dim=-1, keepdim=True)
 
-    q_out = (q_f32 * torch.rsqrt(q_var + eps) * q_weight.to(torch.float32)).to(
-        orig_dtype)
-    k_out = (k_f32 * torch.rsqrt(k_var + eps) * k_weight.to(torch.float32)).to(
-        orig_dtype)
+    q_out = (q_f32 * torch.rsqrt(q_var + eps) * q_weight.to(torch.float32)).to(orig_dtype)
+    k_out = (k_f32 * torch.rsqrt(k_var + eps) * k_weight.to(torch.float32)).to(orig_dtype)
 
     q_3d = q_out.view(q.shape[0], -1, head_dim).contiguous()
     k_3d = k_out.view(k.shape[0], -1, head_dim).contiguous()
-    q_3d, k_3d = _apply_rope_neox(q_3d, k_3d, cos.contiguous(), sin.contiguous(),
-                                  rotary_dim)
+    q_3d, k_3d = _apply_rope_neox(q_3d, k_3d, cos.contiguous(), sin.contiguous(), rotary_dim)
 
     return (
         q_3d.view(q.shape[0], q_hidden_size).contiguous(),
@@ -124,8 +119,9 @@ def _reference_impl(
 @pytest.mark.parametrize("seed", SEEDS)
 @pytest.mark.parametrize("device", DEVICES)
 @torch.inference_mode()
-def test_split_qkv_tp_rmsnorm_rope(num_tokens, num_q_heads, num_kv_heads, head_dim,
-                             rotary_dim, tp_world, eps, dtype, seed, device):
+def test_split_qkv_tp_rmsnorm_rope(
+    num_tokens, num_q_heads, num_kv_heads, head_dim, rotary_dim, tp_world, eps, dtype, seed, device
+):
     torch.manual_seed(seed)
     np.random.seed(seed)
     torch.set_default_device(device)
@@ -134,14 +130,9 @@ def test_split_qkv_tp_rmsnorm_rope(num_tokens, num_q_heads, num_kv_heads, head_d
     q_hidden_size = num_q_heads * head_dim
     kv_hidden_size = num_kv_heads * head_dim
 
-    qkv = torch.randn(num_tokens,
-                      q_hidden_size + kv_hidden_size * 2,
-                      dtype=dtype,
-                      device=device)
-    q_weight = torch.randn(q_hidden_size, dtype=torch.float32,
-                           device=device) * 0.1 + 1.0
-    k_weight = torch.randn(kv_hidden_size, dtype=torch.float32,
-                           device=device) * 0.1 + 1.0
+    qkv = torch.randn(num_tokens, q_hidden_size + kv_hidden_size * 2, dtype=dtype, device=device)
+    q_weight = torch.randn(q_hidden_size, dtype=torch.float32, device=device) * 0.1 + 1.0
+    k_weight = torch.randn(kv_hidden_size, dtype=torch.float32, device=device) * 0.1 + 1.0
     cos, sin = _build_rope(num_tokens, rotary_dim, dtype, device)
 
     q_fused, k_fused, v_fused = _fused_impl(
@@ -171,18 +162,9 @@ def test_split_qkv_tp_rmsnorm_rope(num_tokens, num_q_heads, num_kv_heads, head_d
         sin=sin,
     )
 
-    torch.testing.assert_close(q_fused.to(torch.float32),
-                               q_ref.to(torch.float32),
-                               atol=DEFAULT_ATOL,
-                               rtol=DEFAULT_RTOL)
-    torch.testing.assert_close(k_fused.to(torch.float32),
-                               k_ref.to(torch.float32),
-                               atol=DEFAULT_ATOL,
-                               rtol=DEFAULT_RTOL)
-    torch.testing.assert_close(v_fused.to(torch.float32),
-                               v_ref.to(torch.float32),
-                               atol=DEFAULT_ATOL,
-                               rtol=DEFAULT_RTOL)
+    torch.testing.assert_close(q_fused.to(torch.float32), q_ref.to(torch.float32), atol=DEFAULT_ATOL, rtol=DEFAULT_RTOL)
+    torch.testing.assert_close(k_fused.to(torch.float32), k_ref.to(torch.float32), atol=DEFAULT_ATOL, rtol=DEFAULT_RTOL)
+    torch.testing.assert_close(v_fused.to(torch.float32), v_ref.to(torch.float32), atol=DEFAULT_ATOL, rtol=DEFAULT_RTOL)
 
     gc.collect()
     torch.npu.empty_cache()
